requirements.txt
python-dotenv
facebook-sdk